Reorganize basic tests
