+++ tests/integration/adapter/aliases/test_aliases.py
