+++ tests/integration/adapter/basic/test_adapter_methods.py
@@ -1,0 +1,9 @@
+from dbt.tests.adapter.basic.test_adapter_methods import BaseAdapterMethod
+
+
+class TestBaseAdapterMethod(BaseAdapterMethod):
+    pass
+
+
+class TestBaseCaching(BaseAdapterMethod):
+    pass
