+++ tests/integration/adapter/basic/test_base.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_base import BaseSimpleMaterializations
+
+
+class TestBaseSimpleMaterializations(BaseSimpleMaterializations):
+    pass
