+++ tests/integration/adapter/basic/test_docs_generate.py
