+++ tests/integration/adapter/basic/test_empty.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_empty import BaseEmpty
+
+
+class TestEmpty(BaseEmpty):
+    pass
