+++ tests/integration/adapter/basic/test_ephemeral.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_ephemeral import BaseEphemeral
+
+
+class TestEphemeral(BaseEphemeral):
+    pass
