+++ tests/integration/adapter/basic/test_generic_tests.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_generic_tests import BaseGenericTests
+
+
+class TestGenericTests(BaseGenericTests):
+    pass
