+++ tests/integration/adapter/basic/test_singular_tests.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_singular_tests import BaseSingularTests
+
+
+class TestSingularTests(BaseSingularTests):
+    pass
