+++ tests/integration/adapter/basic/test_singular_tests_ephemeral.py
@@ -1,9 +1,4 @@
-from dbt.tests.adapter.basic.test_singular_tests import BaseSingularTests
 from dbt.tests.adapter.basic.test_singular_tests_ephemeral import BaseSingularTestsEphemeral
-
-
-class TestSingularTests(BaseSingularTests):
-    pass
 
 
 class TestSingularTestsEphemeral(BaseSingularTestsEphemeral):
