+++ tests/integration/adapter/basic/test_snapshot_check_cols.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_snapshot_check_cols import BaseSnapshotCheckCols
+
+
+class TestSnapshotCheckCols(BaseSnapshotCheckCols):
+    pass
