+++ tests/integration/adapter/basic/test_snapshot_timestamp.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_snapshot_timestamp import BaseSnapshotTimestamp
+
+
+class TestSnapshotTimestamp(BaseSnapshotTimestamp):
+    pass
