+++ tests/integration/adapter/basic/test_table_materialization.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_table_materialization import BaseTableMaterialization
+
+
+class TestTableMat(BaseTableMaterialization):
+    pass
