+++ tests/integration/adapter/basic/test_validate_connection.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.basic.test_validate_connection import BaseValidateConnection
+
+
+class TestValidateConnection(BaseValidateConnection):
+    pass
