+++ tests/integration/adapter/incremental/test_base_incremental.py
@@ -1,6 +1,6 @@
 import pytest
 from dbt.tests.adapter.basic.files import model_incremental, schema_base_yml
-from dbt.tests.adapter.basic.test_incremental import BaseIncremental, BaseIncrementalNotSchemaChange
+from dbt.tests.adapter.basic.test_incremental import BaseIncremental
 from dbt.tests.util import run_dbt
 
 uniq_schema = """
@@ -148,19 +148,3 @@ class TestInsertsOnlyIncrementalMaterialization(BaseIncremental):
         }
 
 
-incremental_not_schema_change_sql = """
-{{ config(materialized="incremental", unique_key="user_id_current_time",on_schema_change="sync_all_columns") }}
-select
-    toString(1) || '-' || toString(now64()) as user_id_current_time,
-    {% if is_incremental() %}
-        'thisis18characters' as platform
-    {% else %}
-        'okthisis20characters' as platform
-    {% endif %}
-"""
-
-
-class TestIncrementalNotSchemaChange(BaseIncrementalNotSchemaChange):
-    @pytest.fixture(scope="class")
-    def models(self):
-        return {"incremental_not_schema_change.sql": incremental_not_schema_change_sql}
